% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boloria}
\alias{boloria}
\title{Climate conditions of the boloria chariclea butterfly}
\format{
\subsection{\code{boloria}}{

A data frame with 5 rows and 5 columns
\describe{
\item{year}{Data collected from 1996 - 2013}
\item{site}{Seven sites where measurements were taken}
\item{sex}{sex of butterfly}
\item{WL}{Wing length}
\item{DOY}{Day of year}
\item{Snow}{Snowfall in year in inches}
\item{mayjun}{Diffence in wing length from May to June}
}
}
\source{
\url{https://datadryad.org/stash/dataset/doi:10.5061/dryad.43gt3}
}
\usage{
boloria
}
\description{
Data from a study of wing lengths of butterflies
}
\keyword{datasets}
