% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dd_rounded1}
\alias{dd_rounded1}
\title{dd_rounded1}
\format{
\subsection{\code{dd_rounded1}}{

A data frame with 6 rows and 5 columns:
\describe{
\item{medium}{Medium sized numbers with decimals, not rounded}
\item{words}{characters; rounding will not apply}
\item{big}{large numbers, rounded to the nearest 10}
\item{mixed}{both large and small numbers, rounded to the nearest 10}
\item{small}{small numbers, not rounded}

}
}
}
\source{
\if{html}{\out{
<none>
}}
}
\usage{
dd_rounded1
}
\description{
A version of decimal_data with only numeric columns with mean greater than 10
rounded to the nearest 10. For testing df_apply only.
}
\keyword{datasets}
