% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dd_rounded2}
\alias{dd_rounded2}
\title{dd_rounded2}
\format{
A data frame with 6 rows and 5 columns:
\describe{
\item{medium}{Medium sized numbers with decimals, rounded}
\item{words}{characters; rounding will not apply}
\item{big}{large numbers, no decimals so rounding doesn't do anything}
\item{mixed}{both large and small numbers, rounded}
\item{small}{small numbers, rounded}

}
}
\usage{
dd_rounded2
}
\description{
A version of decimal_data with all numeric columns rounded to 3
deimal places.
}
\keyword{datasets}
