% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{decimal_data}
\alias{decimal_data}
\title{Decimal data}
\format{
A data frame with 6 rows and 5 columns:
\describe{
\item{medium}{Medium sized numbers with decimals}
\item{words}{characters; rounding will not apply}
\item{big}{large numbers}
\item{mixed}{both large and small numbers}
\item{small}{small numbers}

}
}
\usage{
decimal_data
}
\description{
More testing data with numbers of varying lengths
}
\keyword{datasets}
