% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_apply.R
\name{newprojdata303.df_apply}
\alias{newprojdata303.df_apply}
\title{Apply a function to rows in the data frame for which a condition is true}
\usage{
newprojdata303.df_apply(.data, .fun, .filter, ...)
}
\arguments{
\item{.data}{A data frame.}

\item{.fun}{A function that takes a vector and returns it with changes.}

\item{.filter}{A function that takes a vector and returns a bool letting
df_apply know whether to apply .fun}
}
\value{
new_df, a updated data frame where columns where .filter returns
true have been changed by .fun
}
\description{
Apply a function to rows in the data frame for which a condition is true
}
\examples{
df_apply(iris, round, is.numeric, digits = 4)
df_apply(simple_df, toupper, is.character)
}
