% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simple_df}
\alias{simple_df}
\title{Testing data for data 303 project}
\format{
\subsection{\code{simple_df}}{

A data frame with 5 rows and 5 columns
\describe{
\item{X1}{A vector of the letters A-E}
\item{X2}{The words the, cat, in, the, hat}
\item{X3}{Random numbers in decimal format}
\item{X4}{Random integers}
\item{X5}{Names with upper and lower case letters}
}
}
\source{
\if{html}{\out{
<No source>
}}
}
\usage{
simple_df
}
\description{
Simple data containing columns of numeric and character data for use in testing
df_apply function.
}
\keyword{datasets}
