% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simple_up}
\alias{simple_up}
\title{Testing data for data 303 project}
\format{
A data frame with 5 rows and 5 columns
\describe{
\item{X1}{A vector of the letters A-E}
\item{X2}{The words THE, CAT, IN, THE, HAT}
\item{X3}{Random numbers in decimal format}
\item{X4}{Random integers}
\item{X5}{Names with upper case letters}
}
\usage{
simple_up
}
\description{
Simple data containing columns of numeric and character data for use in testing
df_apply function, all in uppercase
}
\keyword{datasets}
